+++ Package.swift
@@ -7,7 +7,9 @@ let package = Package(
     name: "PrintingKit",
     platforms: [
         .iOS(.v13),
-        .macOS(.v11)
+        .macOS(.v11),
+        .tvOS(.v13),
+        .watchOS(.v7)
     ],
     products: [
         .library(