+++ Sources/PrintingKit/PrinterView.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2023 Daniel Saidi. All rights reserved.
 //
 
+#if os(iOS) || os(macOS)
 import SwiftUI
 
 /**
@@ -124,3 +125,4 @@ public extension PrinterView {
         printInTask(view, withScale: scale, printer: printer)
     }
 }
+#endif